pyproject.toml
[project]
name = "verifai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.38.0",
    "langgraph>=0.6.7",
    "python-dotenv>=1.0.0",
    "sentencepiece>=0.1.99",
    "torch>=2.8.0",
    "transformers>=4.56.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "perplexityai>=0.1.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
]
